# docs/openapi.yaml
openapi: 3.0.3
info:
  title: NBE10-12-2-Team1 API
  description: |
    오늘 뭐먹지 SNS 서비스 REST API 문서입니다.

    ## 인증
    인증이 필요한 API는 요청 헤더에 Bearer 토큰을 포함해야 합니다.
    ```
    Authorization: Bearer {accessToken}
    ```
  version: 1.0.0
servers:
  - url: https://api.example.com
    description: Production Server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/v1/users/{id}:
    get:
      tags: [Users]
      summary: 프로필 조회
      description: 사용자 ID로 프로필 정보를 조회합니다. 본인 조회 시 isOwnProfile=true를 반환하며, 타인 조회 시 isOwnProfile=false와 isFollowing 여부를 함께 반환합니다.
      parameters:
        - name: id
          in: path
          required: true
          description: 조회할 사용자 ID
          schema:
            type: integer
            example: 1
      responses:
        "200":
          description: 프로필 조회 성공
          content:
            application/json:
              example:
                id: 1
                nickname: "푸디보이"
                profileImage: "https://k.kakaocdn.net/.../img.jpg"
                email: "foodie@kakao.com"
                provider: "KAKAO"
                createdAt: "2026-06-20T10:00:00"
                isOwnProfile: true
                isFollowing: false
        "401":
          description: 인증 토큰 없음 또는 만료
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "UNAUTHORIZED"
                  message: "인증이 만료되었습니다. 다시 로그인해주세요."
        "404":
          description: 존재하지 않는 사용자 (USER_NOT_FOUND)
          content:
            application/json:
              example:
                status: 404
                message: "User not found: 1"
        "500":
          description: 서버 오류
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INTERNAL_SERVER_ERROR"
                  message: "서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요."
  /api/v1/users/me:
    get:
      tags: [Users]
      summary: 내 프로필 조회
      description: 로그인한 사용자의 프로필 정보를 조회합니다.
      responses:
        "200":
          description: 내 프로필 조회 성공
          content:
            application/json:
              example:
                success: true
                data:
                  id: 1
                  nickname: "푸디보이"
                  profileImage: "https://k.kakaocdn.net/.../img.jpg"
                  email: "foodie@kakao.com"
                  provider: "KAKAO"
                  createdAt: "2026-06-20T10:00:00"
                  isOwnProfile: true
                  isFollowing: false
        "401":
          description: 인증 토큰 없음 또는 만료
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "UNAUTHORIZED"
                  message: "인증이 만료되었습니다. 다시 로그인해주세요."
        "500":
          description: 서버 오류
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INTERNAL_SERVER_ERROR"
                  message: "서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요."

    patch:
      tags: [Users]
      summary: 내 프로필 수정
      description: |
        본인 프로필의 닉네임, 이메일, 비밀번호를 수정합니다.
        LOCAL 사용자는 이메일/비밀번호 변경이 가능하고, KAKAO 사용자는 닉네임만 수정 가능합니다.
        프로필 이미지는 `/api/v1/users/me/image`를 사용합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nickname:
                  type: string
                  minLength: 1
                  maxLength: 20
                  example: "새로운닉네임"
                email:
                  type: string
                  format: email
                  nullable: true
                  example: "new@example.com"
                currentPassword:
                  type: string
                  nullable: true
                  example: "oldPassword123!"
                newPassword:
                  type: string
                  minLength: 8
                  maxLength: 20
                  nullable: true
                  example: "newPassword123!"
      responses:
        "200":
          description: 프로필 수정 성공
          content:
            application/json:
              example:
                success: true
                data:
                  id: 1
                  nickname: "새로운닉네임"
                  profileImage: "https://example.com/profile.jpg"
                  email: "new@example.com"
                  provider: "LOCAL"
                  createdAt: "2026-07-03T10:00:00"
                  isOwnProfile: true
                  isFollowing: false
        "400":
          description: 입력값 오류 / 카카오 사용자 비밀번호 변경 / 현재 비밀번호 불일치
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INVALID_INPUT_VALUE"
                  message: "입력값이 올바르지 않습니다."
        "401":
          description: 인증 토큰 없음 또는 만료
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "UNAUTHORIZED"
                  message: "인증이 만료되었습니다. 다시 로그인해주세요."
        "409":
          description: 이미 사용 중인 닉네임
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "DUPLICATE_NICKNAME"
                  message: "이미 사용 중인 닉네임입니다."
        "500":
          description: 서버 오류
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INTERNAL_SERVER_ERROR"
                  message: "서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요."
  /api/v1/users/me/image:
    patch:
      tags: [Users]
      summary: 프로필 이미지 변경
      description: |
        본인 프로필 이미지를 즉시 업로드하여 변경합니다.
        jpg, jpeg, png 파일만 허용합니다.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  format: binary
                  description: 프로필 이미지 파일 (jpg/jpeg/png)
      responses:
        "200":
          description: 프로필 이미지 변경 성공
          content:
            application/json:
              example:
                success: true
                data:
                  id: 1
                  nickname: "푸디보이"
                  profileImage: "https://example.com/new-profile.jpg"
                  email: "foodie@kakao.com"
                  provider: "KAKAO"
                  createdAt: "2026-06-20T10:00:00"
                  isOwnProfile: true
                  isFollowing: false
        "400":
          description: 이미지 누락 / 허용되지 않는 확장자
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INVALID_INPUT_VALUE"
                  message: "jpg, jpeg, png 파일만 업로드 가능합니다."
        "401":
          description: 인증 토큰 없음 또는 만료
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "UNAUTHORIZED"
                  message: "인증이 만료되었습니다. 다시 로그인해주세요."
        "500":
          description: 서버 오류
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INTERNAL_SERVER_ERROR"
                  message: "서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요."

  # ─────────────────────────────────────────────
  # Follow
  # ─────────────────────────────────────────────
...